Removed endless while loop
If during the 'Set-TargetResource' a new cluster is being created, but the 'New-Cluster' command does not end successfully. Then the 'while' loop was running forever keeping the LCM in a busy state.
I noticed the problem when creating a cluster but the CNO already existed in AD and was linked to a previously created cluster.