Add second node to cluster
When trying to add a second node to an existing cluster in Azure, the ARP restrictions means this fails if we go via the CNO.

To get round this we can query the AD for all the computer objects that are 1) not the CNO 2) have an SPN for MS clustering 3) is not the current node. From there we can query for the clustername (CNO) and determine the owner node.

Using the owner node info we can target that as the CNO name and add the second node.